Arduino doesn't support 8 byte doubles, changing telemetry header for cross platform compatibility

+++ glue_and_networking/lander_protocol.h
@@ -9,11 +9,11 @@
 struct telemetry_packet
 {
 	uint16_t magic;
-	double latitude;
-	double longitude;
+	float latitude;
+	float longitude;
 	uint16_t humidity;
-	double temperature;
-	double altitude;
+	float temperature;
+	float altitude;
 	uint16_t light;	
 };
